+++ packages/react/src/dialog/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @melio-ui/modal
+# @melio-ui/dialog
+
+## 0.0.5
+
+### Patch Changes
+
+- Modal을 Dialog 으로 컴포넌트명 변경
 
 ## 0.0.4